requirements.txt
discord.py~=2.2.0
Columnar~=1.3.1
jishaku~=2.3.2
buttons~=0.1.9
